Add tests for scheduleFrequency and move to its own file
